+++ lib/tbib_gen_timezone_serializable.dart
@@ -1,5 +1,7 @@
 /// A Very Good Project created by Very Good CLI.
 library tbib_gen_timezone_serializable;
 
+export 'package:tbib_timezone_offset/tbib_timezone_offset.dart';
+
 export 'src/ext_in_timezone/ext_timezone.dart';
 export 'src/tbib_gen_timezone_serializable.dart';