+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- now can use JsonTimezoneUTCIsoConverter to send utc time to api
+- and can use JsonTimezoneWithOffsetIsoConverter to send time with offset to api
+- can format date and display it ui with formateDate extension in date time or string
+
 ## 0.0.1
 
 - init